Move logger under services
